refactor(hal): split some dummy modules
