+++ src/ariel-os-hal/src/dummy/identity.rs
@@ -1,0 +1,3 @@
+use ariel_os_embassy_common::identity;
+
+pub type DeviceId = identity::NoDeviceId<identity::NotImplemented>;
