+++ src/ariel-os-hal/src/dummy/peripheral.rs
@@ -1,0 +1,17 @@
+pub use embassy_hal_internal::Peripheral;
+
+/// Dummy type.
+///
+/// See the `OptionalPeripherals` type of your Embassy HAL crate instead.
+#[doc(hidden)]
+pub struct OptionalPeripherals;
+
+/// Dummy type.
+#[doc(hidden)]
+pub struct Peripherals;
+
+impl From<Peripherals> for OptionalPeripherals {
+    fn from(_peripherals: Peripherals) -> Self {
+        Self {}
+    }
+}
